Separate login fragments and service into separate modules
Fragments and services are injected differently - typically the
fragments need only be scoped to a LoginActivity and don't need
to be exposed at the Application level.

Services are different - HasServiceInjector only works when
implemented by Applications, so we're forced to expose the service
to the top-level (App)Component.

This commit puts them into different modules so that an app using dagger
subcomponents has the option to only use LoginServiceModule at the
AppComponent level.