Send redirectUrl to listener and strip protocol before saving
If the url the user logs in with is redirected, we want to know the
redirect url and use that for login. We save this url (or the original
url) to app prefs with the protocol stripped since it's not needed
for the rest of the login process and can cause issues when looking
up the site by the url.