Remove confusion over tag usage by renaming and fix social login bug
These changes will fix an issue where the `LoginEmailFragment` cannot be found upon successful social login
so the app appears to hang when really it just has no way to move forward because the TAG used recall the
fragment from the stack wasn't the same one used in it's creation. This confusion happens because there are
two different layouts now for the `LoginEmailFragment`, one that has an option to login with site creds, and
one that does not.