Refactor: Add back unused got unregistered email login interface method
This 'gotUnregisteredEmail(...)' method is indeed now unused for
the 'WordPress' app, but not for the 'WooCommerce' app. As such, this is
added back as an interface method to be shared with both apps. The only
difference is that for the 'WordPress' app it will not be doing anything
as of now.