Merge WPAndroid changes from '239d8b4701f44993f7648d2c1a1333981998821c'
This is a manual merge from WPAndroid's '239d8b4701f44993f7648d2c1a1333981998821c' hash, which is the current `develop` at the time of this commit. This was done by copying the files from "libs/login" and ignoring the "gradlew.bat" change so that we can update Gradle files with its own script.

This is the first step towards bringing this library in a state where we
can remove subtrees from WPAndroid and WCAndroid clients.