Refactor: Extract or layout
This is done in order for the or layout to be later on easily reused. On
the newly created shared 'login_or_layout' bottom and top padding were
added based on its overall usage. This way, every client that is to be
using this newly created shared 'login_or_layout' will conform to its
now defined standards.

Note: As part of this change the '6dp' login layout
'layout_marginBottom' was replaced by the inner '8dp' or layout
'paddingBottom' and 'paddingTop'. The 'paddingTop' change is irrelevant
since the 'login_email_screen' layout will anyway adjust that
automattically. However, the 'paddingBottom' change will have a diff of
'2dp'. This change can be also considered as negligable, since it is
almost unnoticed.