Build: Extract dependency versions to root build gradle for all modules
This will now make sure that from now on all modules will be pointing to
the same dependency versions, making such upgrades much more
straightforward as that will apply to all modules. This will alleviate
the risk of forgetting to update a specific module, or set of modules
due to an oversight.

FYI: Although this repo has only one module at the moment, this doesn't
mean that another module wouldn't be created in the future, for example
an 'example' app module. As such, this commit is firstly done as a best
practice, then so that all repos start becoming more consistent with
each other and finally as a preparation step for new module creation.