Build: Use double quotes instead of single quotes for dependencies
This is done because:
1. You get the below warning/error if you try to use a syntax like
'androidx.appcompat:appcompat:${version}', note the ${version}:
"It looks like you are trying to substitute a version variable, but
using single quotes ('). For Groovy string interpolation you must use
double quotes (")."
2. If this repo is ever to be migrated from Groovy to Kotlin Gradle DSL,
then this would become a prerequisite for the migration, that is, to
have double quotes (") everywhere.