Analysis: Suppress deprecated warning
Warning Message: "'onActivityCreated(Bundle?): Unit' is deprecated.
Deprecated in Java"

This warning is suppressed, that is, instead of it being resolved, since
a resolution would require a proper investigation and testing. As such,
it might be best to ignore this as out of scope, for now, and so
as to not introduce any breaking changes to this library related module.