CI: Change lint to lint release check
This is done because otherwise both 'lintDebug' and 'lintRelease' will
be run as part of this CI check. This is causing unnecessary overhead,
and thus increasing the overall time it takes for this CI check to
complete.

Having 'lintRelease' is enough. This change also aligns this pipeline
with all the other such project pipelines, and more so CI Lint checks.