Deps: Remove androidx gridlayout dependency
This dependency is completely unused.

FYI: It got introduced as part of this
(f50df670e63312e2dc0973f65e7568a8656f8c95) commit, which updated the
support libs to a newer version (not sure why).

------------------------------------------------------------------------

This removal was suggested by the dependency analysis report, see below:

Advice for :WordPressLoginFlow
Unused dependencies which should be removed:
  implementation 'androidx.gridlayout:gridlayout:1.0.0'
  ...