Deps: Remove androidx legacy support v13 dependency
This dependency is completely unused.

FYI: It got introduced as part of this
(d3dc35035ae5266039dd9d93b359704c375b8fd2) commit, a part of the
AndroidX migration (not sure why).

------------------------------------------------------------------------

This removal was suggested by the dependency analysis report, see below:

Advice for :WordPressLoginFlow
Unused dependencies which should be removed:
  implementation 'androidx.legacy:legacy-support-v13:1.0.0'
  ...