Deps: Remove androidx media dependency
This dependency is completely unused.

FYI: It got introduced as part of this
(f50df670e63312e2dc0973f65e7568a8656f8c95 and then
d3dc35035ae5266039dd9d93b359704c375b8fd2 and) commits, first an update
to the support libs to a newer version and then as part of the AndroidX
migration (not sure why).

------------------------------------------------------------------------

This removal was suggested by the dependency analysis report, see below:

Advice for :WordPressLoginFlow
Unused dependencies which should be removed:
  implementation 'androidx.media:media:1.2.1'
  ...